Beef up testing.
Move basic unit tests into each platform store, where they can test the specific errors for that store.

Add basic integration tests.
